修改检测逻辑 修改临时检测逻辑 精简代码

+++ youtube.py
@@ -17,25 +17,31 @@ class Youtube:
 
     # 关于SearchAPI的文档 https://developers.google.com/youtube/v3/docs/search/list
     def get_videoid_by_channel_id(self):
-        channel_info = json.loads(fetch_html(rf'https://www.googleapis.com/youtube/v3/search?key={self.api_key}'
-                                             rf'&channelId={self.channel_id}&part=snippet,id&order=date&maxResults=5'))
+        channel_info = json.loads(fetch_html(rf'https://www.googleapis.com/youtube/v3/search?part=snippet&'
+                                             rf'channelId={self.channel_id}&eventType=live&maxResults=1&type=video&'
+                                             rf'key={self.api_key}'))
         # 判断获取的数据是否正确
         if channel_info['items']:
-            item = channel_info['items']
-            vid = [x['id']['videoId'] for x in item]
-            return vid
+            item = channel_info['items'][0]
+            title = item['snippet']['title']
+            vid = item['id']['videoId']
+            return {'Title': title,
+                    'Ref': vid}
 
     def get_temp_refvid(self, temp_ref):
         reg = r"watch\?v=([A-Za-z0-9_-]{11})"
         idre = re.compile(reg)
         for _id, _ref in temp_ref:
-            vid = [re.search(idre, _ref).group(1)]
-            is_live = self.getlive_vid(vid)
-            if is_live:
-                self.database.delete(_id)
-                return is_live
+            vid = re.search(idre, _ref).group(1)
+            html = fetch_html("https://www.youtube.com/watch?v=" f"{vid}")
+            if '"isLive":true' in html:
+                is_live = self.getlive_title([vid])
+                # is_live = self.getlive_vid(vid)
+                if is_live:
+                    self.database.delete(_id)
+                    return is_live
 
-    def getlive_vid(self, vid):
+    def getlive_title(self, vid):
         for x in vid:
             live_info = json.loads(
                 fetch_html(rf'https://www.googleapis.com/youtube/v3/videos?id={x}&key={self.api_key}&'
@@ -45,24 +51,16 @@ class Youtube:
                 raise ValueError
             # JSON中的数组将被转换为列表，此处使用[0]获得其中的数据
             item = live_info['items'][0]
-            snippet = item['snippet']
-            info_dict = {'Title': snippet['title'],
-                         'Islive': snippet['liveBroadcastContent']}
-            if info_dict.get('Islive') == 'live':
-                return {'Title': info_dict['Title'],
-                        'Ref': x}
-            elif info_dict.get('Islive') == 'upcoming':
-                echo_log('Youtube' + time.strftime('|%m-%d %H:%M:%S|', time.localtime(time.time())) +
-                         f'{info_dict["Title"]} is upcoming, waiting to recheck')
-            else:
-                echo_log('Youtube' + time.strftime('|%m-%d %H:%M:%S|', time.localtime(time.time())) +
-                         f'{info_dict["Title"]} is not a live video')
+            title = item['snippet']['title']
+            return {'Title': title,
+                    'Ref': x}
 
     def check(self):
         html = fetch_html(f'https://www.youtube.com/channel/{self.channel_id}/featured')
-        if 'LIVE NOW' in html:
-            vid = self.get_videoid_by_channel_id()
-            is_live = self.getlive_vid(vid)
+        if '"label":"LIVE NOW"' in html:
+            # vid = self.get_videoid_by_channel_id()
+            # is_live = self.getlive_vid(vid)
+            is_live = self.get_videoid_by_channel_id()
             process_video(is_live, 'Youtube')
         elif 'Upcoming live streams' in html:
             echo_log('Youtube' + time.strftime('|%m-%d %H:%M:%S|', time.localtime(time.time())) +
